feat: updated UI. Created button disable feature.
